Made my python version
It works with 3 inputs isKeySmash(string, int repetition, int vowel)

string: the string to be evaluated

int repetition: integer with the maximum percentage of repeated letters that is considered KeySmash (33% seems to work fine) If theres more than 33% of repeated letters in a word its KeySmash.

int vowel:  integer with the maximum percentage of vowels letters that is considered KeySmash (25% - 33% seems to work fine). If theres less than 33% of vowels in a word its probably a KeySmash, but that´s not always true since Vowels are distributes evenly across the keyboard